feat : adding cmake and moving folders
