+++ run.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+
+cd build
+wait
+cmake ..
+wait
+make
+./SGH-LP1
