+++ src/Patient.cpp
@@ -8,7 +8,8 @@
 //         this->fixDate();
 //     }
 
-
-// // ------------------------------------------------------------------------------
+// //
+// ------------------------------------------------------------------------------
 // // Métodos:
-// // ------------------------------------------------------------------------------
+// //
+// ------------------------------------------------------------------------------
